# limedaichatserver/src/docs/openapi.yml
openapi: 3.0.0
info:
  title: LiMedAi API Documentation
  version: 1.0.1
  description: API documentation for LiMedAi server for refrence
host: http://localhost:xxxx
basePath: /
servers:
  - url: http://localhost:xxxx
    description: Development server example
paths:
  /api/authentication/signup:
    post:
      summary: Register a new user
      description: >
        Creates a new user account with the provided email and password. On
        successful registration, it sets a JWT cookie for authentication and
        returns the user's basic information.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: 用户@example.com
                password:
                  type: string
                  format: password
                  example: 强密码123
      responses:
        '201':
          description: User successfully created
          headers:
            Set-Cookie:
              schema:
                type: string
                example: jwt=abcde12345; Path=/; HttpOnly
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Bad request - email or password missing
          content:
            text/plain:
              schema:
                type: string
                example: 需要电子邮件和密码
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: 服务器内部错误
  /api/authentication/signin:
    post:
      summary: Authenticate a user and retrieve their information
      description: >
        Authenticates a user with the provided email and password. On successful
        authentication, it sets a JWT cookie for session management and returns
        the user's detailed information.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: 用户@example.com
                password:
                  type: string
                  format: password
                  example: 密码123
      responses:
        '200':
          description: User successfully authenticated
          headers:
            Set-Cookie:
              schema:
                type: string
                example: jwt=abcde12345; Path=/; HttpOnly
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserDetailed'
        '400':
          description: Bad request - email or password missing, or authentication failed
          content:
            text/plain:
              schema:
                type: string
                example: 需要电子邮件和密码
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: 服务器内部错误
  /api/authentication/fetchuserinfo:
    get:
      summary: Fetch authenticated user's information
      description: >
        Retrieves detailed information for the currently authenticated user.
        This endpoint requires a valid JWT token in the cookie for
        authentication.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User information successfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserDetailed'
        '401':
          description: Unauthorized - No valid JWT token provided
          content:
            text/plain:
              schema:
                type: string
                example: 您未通过身份验证
        '403':
          description: Forbidden - Invalid JWT token
          content:
            text/plain:
              schema:
                type: string
                example: 无效的令牌
        '404':
          description: User not found
          content:
            text/plain:
              schema:
                type: string
                example: 未找到用户
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: 服务器内部错误
  /api/profile/updateprofile:
    patch:
      summary: Update the user's profile information
      description: >
        The JWT token in the 'jwt' cookie is verified by the verifyJWT
        middleware. If valid, the user's ID is extracted from the token and
        added to request.body.userId. The updateProfile controller then updates
        the user's profile information and returns the updated user data.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  example: 约翰
                lastName:
                  type: string
                  example: 多伊
                theme:
                  type: number
                  example: 1
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 60d0fe4f5311236168a109ca
                      email:
                        type: string
                        example: user@example.com
                      configuredProfile:
                        type: boolean
                        example: true
                      firstName:
                        type: string
                        example: 约翰
                      lastName:
                        type: string
                        example: 多伊
                      avatar:
                        type: string
                        example: https://example.com/avatar.jpg
                      theme:
                        type: number
                        example: 1
        '400':
          description: Bad Request, missing required fields
          content:
            text/plain:
              schema:
                type: string
                example: 名字、姓氏和主题均为必填项
        '401':
          description: Unauthorized, JWT cookie is missing
          content:
            text/plain:
              schema:
                type: string
                example: 您未通过身份验证
        '403':
          description: Forbidden, JWT token is invalid
          content:
            text/plain:
              schema:
                type: string
                example: 无效的令牌
        '404':
          description: User not found
          content:
            text/plain:
              schema:
                type: string
                example: 错误
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: 服务器内部错误
  /api/profile/updateavatar:
    patch:
      summary: Update the user's avatar image
      description: >
        The JWT token in the 'jwt' cookie is verified by the verifyJWT
        middleware. If valid, the user's ID is extracted from the token and
        added to request.body.userId. The uploadImage middleware then handles
        the file upload. Finally, the updateAvatar controller updates the user's
        avatar and returns the response.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
      responses:
        '200':
          description: Avatar updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  avatar:
                    type: string
                    example: src/assets/uploads/avatars/1726702800641EcoNex_A10.jpg
        '400':
          description: Bad Request, no file uploaded
          content:
            text/plain:
              schema:
                type: string
                example: 需要上传图片
        '401':
          description: Unauthorized, JWT cookie is missing
          content:
            text/plain:
              schema:
                type: string
                example: 您未通过身份验证
        '403':
          description: Forbidden, JWT token is invalid
          content:
            text/plain:
              schema:
                type: string
                example: 无效的令牌
        '404':
          description: User not found
          content:
            text/plain:
              schema:
                type: string
                example: 未找到用户
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: 服务器内部错误
  /api/profile/deleteavatar:
    delete:
      summary: Delete the user's avatar image
      description: >
        The JWT token in the 'jwt' cookie is verified by the verifyJWT
        middleware. If valid, the user's ID is extracted from the token and
        added to request.body.userId. The deleteAvatar controller then handles
        deleting the user's avatar and returns the response.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Avatar deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 用户头像删除成功
                  deleted:
                    type: boolean
                    example: true
        '401':
          description: Unauthorized, JWT cookie is missing
          content:
            text/plain:
              schema:
                type: string
                example: 您未通过身份验证
        '403':
          description: Forbidden, JWT token is invalid
          content:
            text/plain:
              schema:
                type: string
                example: 无效的令牌
        '404':
          description: User not found
          content:
            text/plain:
              schema:
                type: string
                example: 未找到用户错误
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: 服务器内部错误
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: 60d0fe4f5311236168a109ca
        email:
          type: string
          format: email
          example: 用户@example.com
        configuredProfile:
          type: boolean
          example: false
    UserDetailed:
      type: object
      properties:
        id:
          type: string
          example: 60d0fe4f5311236168a109ca
        email:
          type: string
          format: email
          example: 用户@example.com
        configuredProfile:
          type: boolean
          example: true
        firstName:
          type: string
          example: 张
        lastName:
          type: string
          example: 三
        avatar:
          type: string
          example: https://example.com/avatar.jpg
        theme:
          type: number
          example: 1
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
tags: []